Added suport for eBPF classification
Now the Classifier has two different configuration types: Standard
and eBPF. Standard mode consists on classification using 5-tuple,
as it was already done before. This is the default method. The
second is a mode in which the user can load an ELF with eBPF code
to be loaded into a userspace eBPF VM running in the Classifier.
This allows the user to configure any kind of classification s/he
might want, allowing greater flexibility.